refactor(app-shell): layoutNodesSection 七组 options 外移为 builder（2531→2182）
