+++ src/app-shell/sections/layoutNodes/layoutNodesOptionsChrome.ts
@@ -1,0 +1,242 @@
+import type { ChromeLayoutNodesOptions } from "../../../features/layout/hooks/layoutNodesTypes";
+
+// B4 三期：useLayoutNodes 的 chrome options 对象逐字外移；键名/语句/注释保持原样，
+// hook 局部变量统一经 bag 传入（引用名不改）。
+export function buildLayoutNodesChromeOptions(
+  bag: Record<string, any>,
+): ChromeLayoutNodesOptions {
+  const {
+    handleOpenSettings,
+    handleOpenSessionManagement,
+    handleOpenShortcutsSettings,
+    handleOpenAgentSettings,
+    handleOpenPromptSettings,
+    handleOpenModelSettings,
+    handleOpenCliSettings,
+    handleRefreshModelConfig,
+    isModelConfigRefreshing,
+    handleOpenSkillsSettings,
+    handleDebugClick,
+    showDebugButton,
+    handleAddWorkspace,
+    handleSelectHome,
+    handleSelectWorkspace,
+    handleReorderWorkspaces,
+    handleConnectWorkspace,
+    handleAddAgent,
+    availableEngines,
+    refreshEngines,
+    handleStartSharedConversation,
+    handleAddWorktreeAgent,
+    handleAddCloneAgent,
+    handleToggleWorkspaceCollapse,
+    handleSelectThread,
+    handleClearActiveThread,
+    handleProviderContinuationTargetReady,
+    handleSelectHomeWorkspace,
+    handleDeleteThread,
+    handleArchiveThread,
+    deleteThreadPrompt,
+    isDeleteThreadPromptBusy,
+    handleDeleteThreadPromptCancel,
+    handleConfirmDeleteConfirm,
+    renamePrompt,
+    handleRenamePromptChange,
+    handleRenamePromptCancel,
+    handleRenamePromptConfirm,
+    handleSyncThread,
+    pinThread,
+    unpinThread,
+    isThreadPinned,
+    getPinTimestamp,
+    pinnedThreadsVersion,
+    isThreadAutoNaming,
+    handleSidebarRenameThread,
+    handleAutoNameThread,
+    handleOpenClaudeTui,
+    handleDeleteWorkspace,
+    handleDeleteWorktree,
+    handleRenameWorkspaceAlias,
+    workspaceGroups,
+    assignWorkspaceGroup,
+    handleLoadOlderThreads,
+    handleQuickReloadWorkspaceThreads,
+    handleReloadWorkspaceThreads,
+    isExitedSessionsHidden,
+    toggleExitedSessionsHidden,
+    rootSessionFolderDraftRequestByWorkspaceId,
+    onRequestRootSessionFolderDraft,
+    updaterState,
+    startUpdate,
+    dismissUpdate,
+    errorToasts,
+    dismissErrorToast,
+    handleOpenSpecHub,
+    showLoadingProgressDialog,
+    hideLoadingProgressDialog,
+    activeWorkspace,
+    activeParentWorkspace,
+    worktreeLabel,
+    worktreeRename,
+    isWorktreeWorkspace,
+    gitStatus,
+    branches,
+    branchError,
+    currentBranch,
+    localBranches,
+    remoteBranches,
+    repositories,
+    repositoriesLoading,
+    repositoryError,
+    selectedRepositoryRoot,
+    selectRepository,
+    handleCheckoutBranch,
+    handleCreateBranch,
+    handleUpdateBranch,
+    handleUpdateAllRepositories,
+    handleCheckoutAllRepositories,
+    handleLoadCommonRepositoryBranches,
+    handleCopyThread,
+    handleLockPanel,
+    handleToggleTerminalPanel,
+    isCompact,
+    launchScriptState,
+    launchScriptsState,
+    mainHeaderActions,
+    filePanelMode,
+    setFilePanelMode,
+    liveEditPreviewEnabled,
+    handleToggleLiveEditPreview,
+    isFilesLoading,
+    fileTreeLoadError,
+    refreshFiles,
+    handleOpenDetachedFileExplorer,
+    handleToggleRuntimeConsole,
+    runtimeRunState,
+    browserDockOpen,
+    handleCloseBrowserDock,
+  } = bag;
+  return {
+    onOpenSettings: handleOpenSettings,
+    onOpenSessionManagement: handleOpenSessionManagement,
+    onOpenShortcutsSettings: handleOpenShortcutsSettings,
+    onOpenAgentSettings: handleOpenAgentSettings,
+    onOpenPromptSettings: handleOpenPromptSettings,
+    onOpenModelSettings: handleOpenModelSettings,
+    onOpenCliSettings: handleOpenCliSettings,
+    onRefreshModelConfig: handleRefreshModelConfig,
+    isModelConfigRefreshing,
+    onOpenSkillsSettings: handleOpenSkillsSettings,
+    onOpenDebug: handleDebugClick,
+    showDebugButton,
+    onAddWorkspace: handleAddWorkspace,
+    onSelectHome: handleSelectHome,
+    onSelectWorkspace: handleSelectWorkspace,
+    onReorderWorkspaces: handleReorderWorkspaces,
+    onConnectWorkspace: handleConnectWorkspace,
+    onAddAgent: handleAddAgent,
+    engineOptions: availableEngines,
+    onRefreshEngineOptions: refreshEngines,
+    onAddSharedAgent: handleStartSharedConversation,
+    onAddWorktreeAgent: handleAddWorktreeAgent,
+    onAddCloneAgent: handleAddCloneAgent,
+    onToggleWorkspaceCollapse: handleToggleWorkspaceCollapse,
+    onSelectThread: handleSelectThread,
+    onClearActiveThread: handleClearActiveThread,
+    onProviderContinuationTargetReady: handleProviderContinuationTargetReady,
+    onSelectHomeWorkspace: handleSelectHomeWorkspace,
+    onDeleteThread: handleDeleteThread,
+    onArchiveThread: handleArchiveThread,
+    deleteConfirmThreadId: deleteThreadPrompt?.threadId ?? null,
+    deleteConfirmWorkspaceId: deleteThreadPrompt?.workspaceId ?? null,
+    deleteConfirmBusy: isDeleteThreadPromptBusy,
+    onCancelDeleteConfirm: handleDeleteThreadPromptCancel,
+    onConfirmDeleteConfirm: handleConfirmDeleteConfirm,
+    renameThreadId: renamePrompt?.threadId ?? null,
+    renameWorkspaceId: renamePrompt?.workspaceId ?? null,
+    renameName: renamePrompt?.name ?? "",
+    onRenameChange: handleRenamePromptChange,
+    onRenameCancel: handleRenamePromptCancel,
+    onRenameConfirm: handleRenamePromptConfirm,
+    onSyncThread: handleSyncThread,
+    pinThread,
+    unpinThread,
+    isThreadPinned,
+    getPinTimestamp,
+    pinnedThreadsVersion,
+    isThreadAutoNaming,
+    onRenameThread: handleSidebarRenameThread,
+    onAutoNameThread: handleAutoNameThread,
+    onOpenClaudeTui: handleOpenClaudeTui,
+    onDeleteWorkspace: handleDeleteWorkspace,
+    onDeleteWorktree: handleDeleteWorktree,
+    onRenameWorkspaceAlias: handleRenameWorkspaceAlias,
+    workspaceGroups,
+    onAssignWorkspaceGroup: assignWorkspaceGroup,
+    onLoadOlderThreads: handleLoadOlderThreads,
+    onQuickReloadWorkspaceThreads: handleQuickReloadWorkspaceThreads,
+    onReloadWorkspaceThreads: handleReloadWorkspaceThreads,
+    isExitedSessionsHidden,
+    onToggleExitedSessionsHidden: toggleExitedSessionsHidden,
+    rootSessionFolderDraftRequestByWorkspaceId,
+    onRequestRootSessionFolderDraft,
+    updaterState,
+    onUpdate: startUpdate,
+    onDismissUpdate: dismissUpdate,
+    errorToasts,
+    onDismissErrorToast: dismissErrorToast,
+    onOpenSpecHub: handleOpenSpecHub,
+    showLoadingProgressDialog,
+    hideLoadingProgressDialog,
+    activeWorkspace,
+    activeParentWorkspace,
+    worktreeLabel,
+    worktreeRename: worktreeRename ?? undefined,
+    isWorktreeWorkspace,
+    branchName: gitStatus.branchName,
+    branches,
+    branchError,
+    branchCurrentName: currentBranch,
+    branchLocalItems: localBranches,
+    branchRemoteItems: remoteBranches,
+    gitRepositories: repositories,
+    gitRepositoriesLoading: repositoriesLoading,
+    gitRepositoriesError: repositoryError,
+    selectedGitRepositoryRoot: selectedRepositoryRoot,
+    onSelectGitRepository: selectRepository,
+    onCheckoutBranch: handleCheckoutBranch,
+    onCreateBranch: handleCreateBranch,
+    onUpdateBranch: handleUpdateBranch,
+    onUpdateAllRepositories: handleUpdateAllRepositories,
+    onCheckoutAllRepositories: handleCheckoutAllRepositories,
+    onLoadCommonRepositoryBranches: handleLoadCommonRepositoryBranches,
+    onCopyThread: handleCopyThread,
+    onLockPanel: handleLockPanel,
+    onToggleTerminal: handleToggleTerminalPanel,
+    showTerminalButton: !isCompact,
+    launchScript: launchScriptState.launchScript,
+    launchScriptEditorOpen: launchScriptState.editorOpen,
+    launchScriptDraft: launchScriptState.draftScript,
+    launchScriptSaving: launchScriptState.isSaving,
+    launchScriptError: launchScriptState.error,
+    onRunLaunchScript: launchScriptState.onRunLaunchScript,
+    onOpenLaunchScriptEditor: launchScriptState.onOpenEditor,
+    onCloseLaunchScriptEditor: launchScriptState.onCloseEditor,
+    onLaunchScriptDraftChange: launchScriptState.onDraftScriptChange,
+    onSaveLaunchScript: launchScriptState.onSaveLaunchScript,
+    launchScriptsState,
+    mainHeaderActions,
+    filePanelMode,
+    onFilePanelModeChange: setFilePanelMode,
+    liveEditPreviewEnabled,
+    onToggleLiveEditPreview: handleToggleLiveEditPreview,
+    fileTreeLoading: isFilesLoading,
+    fileTreeLoadError,
+    onRefreshFiles: refreshFiles,
+    onOpenDetachedFileExplorer: handleOpenDetachedFileExplorer,
+    onToggleRuntimeConsole: handleToggleRuntimeConsole,
+    runtimeConsoleVisible: runtimeRunState.runtimeConsoleVisible,
+    browserDockOpen,
+    onCloseBrowserDock: handleCloseBrowserDock,
+  };
+}
