+++ src/app-shell/sections/layoutNodes/layoutNodesOptionsComposer.ts
@@ -1,0 +1,267 @@
+import type { ComposerLayoutNodesOptions } from "../../../features/layout/hooks/layoutNodesTypes";
+
+// B4 三期：useLayoutNodes 的 composer options 对象逐字外移；键名/语句/注释保持原样，
+// hook 局部变量统一经 bag 传入（引用名不改）。
+export function buildLayoutNodesComposerOptions(
+  bag: Record<string, any>,
+): ComposerLayoutNodesOptions {
+  const {
+    handleSendPrompt,
+    handleSendPromptToNewAgent,
+    handleCreatePrompt,
+    handleUpdatePrompt,
+    handleDeletePrompt,
+    handleMovePrompt,
+    handleRevealWorkspacePrompts,
+    handleRevealGeneralPrompts,
+    activeWorkspace,
+    handleComposerSendWithIntentCanvas,
+    handleComposerQueueWithIntentCanvas,
+    handleRequestContextCompaction,
+    interruptTurn,
+    activeThreadId,
+    isSharedSessionThreadId,
+    completionEmailIntentByThread,
+    handleToggleCompletionEmail,
+    handleRewindFromMessage,
+    handleForkFromMessage,
+    canInterrupt,
+    isReviewing,
+    isProcessing,
+    appSettings,
+    reviewPrompt,
+    closeReviewPrompt,
+    showPresetStep,
+    choosePreset,
+    highlightedPresetIndex,
+    setHighlightedPresetIndex,
+    highlightedBranchIndex,
+    setHighlightedBranchIndex,
+    highlightedCommitIndex,
+    setHighlightedCommitIndex,
+    handleReviewPromptKeyDown,
+    selectBranch,
+    selectBranchAtIndex,
+    confirmBranch,
+    selectCommit,
+    selectCommitAtIndex,
+    confirmCommit,
+    updateCustomInstructions,
+    confirmCustom,
+    activeTokenUsage,
+    activeEngine,
+    handleCodexAutoCompactionSettingsChange,
+    activeQueue,
+    handleDraftChange,
+    activeImages,
+    pickImages,
+    attachImages,
+    removeImage,
+    prefillDraft,
+    handlePrefillHandled,
+    composerInsert,
+    handleInsertHandled,
+    handleEditQueued,
+    handleDeleteQueued,
+    handleFuseQueued,
+    canFuseActiveQueue,
+    fuseDisabledReasonKey,
+    activeFusingMessageId,
+    collaborationModes,
+    collaborationModesEnabled,
+    selectedCollaborationModeId,
+    applySelectedCollaborationMode,
+    availableEngines,
+    handleSelectConversationEngine,
+    effectiveModels,
+    providerModelCatalogs,
+    effectiveSelectedModelId,
+    projectMapDatasetController,
+    handleSelectModel,
+    intentCanvasOpenRequest,
+    handleOpenIntentCanvas,
+    handleIntentCanvasOpenRequestConsumed,
+    handleAttachIntentCanvasToThread,
+    pendingIntentCanvasDocuments,
+    handleRemovePendingIntentCanvas,
+    reasoningOptions,
+    selectedEffort,
+    setSelectedEffort,
+    claudeThinkingVisible,
+    handleResolvedClaudeThinkingVisibleChange,
+    effectiveReasoningSupported,
+    openCodeAgents,
+    selectedOpenCodeAgent,
+    handleSelectOpenCodeAgent,
+    selectedAgent,
+    handleSelectAgent,
+    OPENCODE_VARIANT_OPTIONS,
+    selectedOpenCodeVariant,
+    handleSelectOpenCodeVariant,
+    accessMode,
+    handleSetAccessMode,
+    skills,
+    EMPTY_STRING_ARRAY,
+    prompts,
+    commands,
+    files,
+    directories,
+    directoryMetadata,
+    fileTreeSourceVersion,
+    gitignoredFiles,
+    gitignoredDirectories,
+    handleInsertComposerText,
+    composerInputRef,
+    composerEditorSettings,
+    formatShortcutLabelOrNull,
+    textareaHeight,
+    onTextareaHeightChange,
+    handleOpenExperimentalSettings,
+    composerSendLabel,
+    activeEditorFilePath,
+    activeEditorLineRange,
+    activeIntentCanvasCodeSelectionAnchor,
+    setActiveIntentCanvasCodeSelectionAnchor,
+    fileReferenceMode,
+    setFileReferenceMode,
+  } = bag;
+  return {
+    onSendPrompt: handleSendPrompt,
+    onSendPromptToNewAgent: handleSendPromptToNewAgent,
+    onCreatePrompt: handleCreatePrompt,
+    onUpdatePrompt: handleUpdatePrompt,
+    onDeletePrompt: handleDeletePrompt,
+    onMovePrompt: handleMovePrompt,
+    onRevealWorkspacePrompts: handleRevealWorkspacePrompts,
+    onRevealGeneralPrompts: handleRevealGeneralPrompts,
+    canRevealGeneralPrompts: Boolean(activeWorkspace),
+    onSend: handleComposerSendWithIntentCanvas,
+    onQueue: handleComposerQueueWithIntentCanvas,
+    onRequestContextCompaction: handleRequestContextCompaction,
+    onStop: interruptTurn,
+    // Shared CLI：完成邮件与 Native turn lifecycle 脱节，隐藏入口避免假能力。
+    // ContextBar 以 onToggleCompletionEmail 是否传入决定是否渲染邮件图标。
+    completionEmailSelected: Boolean(
+      activeThreadId &&
+      !isSharedSessionThreadId(activeThreadId) &&
+      completionEmailIntentByThread?.[activeThreadId],
+    ),
+    completionEmailDisabled:
+      !activeThreadId || isSharedSessionThreadId(activeThreadId),
+    onToggleCompletionEmail: isSharedSessionThreadId(activeThreadId)
+      ? undefined
+      : handleToggleCompletionEmail,
+    onRewind: handleRewindFromMessage,
+    onForkFromMessage: handleForkFromMessage,
+    canStop: canInterrupt,
+    isReviewing,
+    isProcessing,
+    steerEnabled: appSettings.experimentalSteerEnabled,
+    reviewPrompt,
+    onReviewPromptClose: closeReviewPrompt,
+    onReviewPromptShowPreset: showPresetStep,
+    onReviewPromptChoosePreset: choosePreset,
+    highlightedPresetIndex,
+    onReviewPromptHighlightPreset: setHighlightedPresetIndex,
+    highlightedBranchIndex,
+    onReviewPromptHighlightBranch: setHighlightedBranchIndex,
+    highlightedCommitIndex,
+    onReviewPromptHighlightCommit: setHighlightedCommitIndex,
+    onReviewPromptKeyDown: handleReviewPromptKeyDown,
+    onReviewPromptSelectBranch: selectBranch,
+    onReviewPromptSelectBranchAtIndex: selectBranchAtIndex,
+    onReviewPromptConfirmBranch: confirmBranch,
+    onReviewPromptSelectCommit: selectCommit,
+    onReviewPromptSelectCommitAtIndex: selectCommitAtIndex,
+    onReviewPromptConfirmCommit: confirmCommit,
+    onReviewPromptUpdateCustomInstructions: updateCustomInstructions,
+    onReviewPromptConfirmCustom: confirmCustom,
+    activeTokenUsage,
+    contextDualViewEnabled: activeEngine === "codex",
+    codexAutoCompactionEnabled: appSettings.codexAutoCompactionEnabled,
+    codexAutoCompactionThresholdPercent:
+      appSettings.codexAutoCompactionThresholdPercent,
+    onCodexAutoCompactionSettingsChange:
+      handleCodexAutoCompactionSettingsChange,
+    activeQueue,
+    onDraftChange: handleDraftChange,
+    activeImages,
+    onPickImages: pickImages,
+    onAttachImages: attachImages,
+    onRemoveImage: removeImage,
+    prefillDraft,
+    onPrefillHandled: handlePrefillHandled,
+    insertText: composerInsert,
+    onInsertHandled: handleInsertHandled,
+    onEditQueued: handleEditQueued,
+    onDeleteQueued: handleDeleteQueued,
+    onFuseQueued: handleFuseQueued,
+    canFuseActiveQueue,
+    fuseDisabledReasonKey,
+    activeFusingMessageId,
+    collaborationModes,
+    collaborationModesEnabled,
+    selectedCollaborationModeId,
+    onSelectCollaborationMode: applySelectedCollaborationMode,
+    engines: availableEngines,
+    selectedEngine: activeEngine,
+    usePresentationProfile: appSettings.chatCanvasUsePresentationProfile,
+    onSelectEngine: handleSelectConversationEngine,
+    models: effectiveModels,
+    providerModelCatalogs,
+    selectedModelId: effectiveSelectedModelId,
+    projectMapDatasetController,
+    onSelectModel: handleSelectModel,
+    intentCanvasOpenRequest,
+    onOpenIntentCanvas: handleOpenIntentCanvas,
+    onIntentCanvasOpenRequestConsumed: handleIntentCanvasOpenRequestConsumed,
+    onAttachIntentCanvasToThread: handleAttachIntentCanvasToThread,
+    pendingIntentCanvasDocuments,
+    onRemovePendingIntentCanvas: handleRemovePendingIntentCanvas,
+    reasoningOptions,
+    selectedEffort,
+    onSelectEffort: setSelectedEffort,
+    claudeThinkingVisible,
+    onResolvedClaudeThinkingVisibleChange:
+      handleResolvedClaudeThinkingVisibleChange,
+    reasoningSupported: effectiveReasoningSupported,
+    opencodeAgents: openCodeAgents,
+    selectedOpenCodeAgent,
+    onSelectOpenCodeAgent: handleSelectOpenCodeAgent,
+    selectedAgent,
+    onSelectAgent: handleSelectAgent,
+    opencodeVariantOptions: OPENCODE_VARIANT_OPTIONS,
+    selectedOpenCodeVariant,
+    onSelectOpenCodeVariant: handleSelectOpenCodeVariant,
+    accessMode,
+    onSelectAccessMode: handleSetAccessMode,
+    skills,
+    customSkillDirectories:
+      appSettings.customSkillDirectories ?? EMPTY_STRING_ARRAY,
+    prompts,
+    commands,
+    files,
+    directories,
+    directoryMetadata,
+    fileTreeSourceVersion,
+    gitignoredFiles,
+    gitignoredDirectories,
+    onInsertComposerText: handleInsertComposerText,
+    textareaRef: composerInputRef,
+    composerEditorSettings,
+    composerSendShortcut: appSettings.composerSendShortcut,
+    composerInterruptShortcutLabel: formatShortcutLabelOrNull(
+      appSettings.interruptShortcut,
+    ),
+    textareaHeight,
+    onTextareaHeightChange,
+    onOpenExperimentalSettings: handleOpenExperimentalSettings,
+    composerSendLabel,
+    activeComposerFilePath: activeEditorFilePath,
+    activeComposerFileLineRange: activeEditorLineRange,
+    activeCodeSelectionAnchor: activeIntentCanvasCodeSelectionAnchor,
+    onActiveCodeSelectionAnchorChange: setActiveIntentCanvasCodeSelectionAnchor,
+    fileReferenceMode,
+    onFileReferenceModeChange: setFileReferenceMode,
+  };
+}
