+++ src/app-shell/sections/layoutNodes/layoutNodesOptionsEditor.ts
@@ -1,0 +1,106 @@
+import type { EditorLayoutNodesOptions } from "../../../features/layout/hooks/layoutNodesTypes";
+
+// B4 三期：useLayoutNodes 的 editor options 对象逐字外移；键名/语句/注释保持原样，
+// hook 局部变量统一经 bag 传入（引用名不改）。
+export function buildLayoutNodesEditorOptions(
+  bag: Record<string, any>,
+): EditorLayoutNodesOptions {
+  const {
+    centerMode,
+    setCenterMode,
+    fileCompareSession,
+    handleOpenFileHistory,
+    editorSplitCompanion,
+    setEditorSplitCompanion,
+    editorSplitLayout,
+    handleToggleEditorSplitLayout,
+    isEditorFileMaximized,
+    handleToggleEditorFileMaximized,
+    activeEditorFilePath,
+    editorNavigationTarget,
+    editorHighlightTarget,
+    openFileTabs,
+    handleActivateWorkspaceFileTab,
+    handleCloseWorkspaceFileTab,
+    handleCloseOtherWorkspaceFileTabs,
+    handleCloseAllWorkspaceFileTabs,
+    handleReorderWorkspaceFileTabs,
+    setActiveEditorLineRange,
+    handleOpenWorkspaceFile,
+    handleOpenWorkspaceFileCompare,
+    handleCloseFileCompare,
+    enableMainFileExternalChangeMonitoring,
+    mainFileExternalChangeTransportMode,
+    liveEditPreviewEnabled,
+    handleExitWorkspaceEditor,
+    handleExitDiff,
+    activeTab,
+    setActiveTab,
+    tabletTab,
+    gitPanelMode,
+    handleGitPanelModeChange,
+    handleOpenGitHistoryPanel,
+    handleOpenProjectMap,
+    gitDiffViewStyle,
+    gitDiffListView,
+    setGitDiffListView,
+    t,
+    activeParentWorkspace,
+    isWorktreeWorkspace,
+    worktreeApplyLoading,
+    worktreeApplyError,
+    worktreeApplySuccess,
+    handleApplyWorktreeChanges,
+  } = bag;
+  return {
+    centerMode,
+    setCenterMode,
+    fileCompareSession,
+    onOpenFileHistory: handleOpenFileHistory,
+    editorSplitCompanion,
+    setEditorSplitCompanion,
+    editorSplitLayout,
+    onToggleEditorSplitLayout: handleToggleEditorSplitLayout,
+    isEditorFileMaximized,
+    onToggleEditorFileMaximized: handleToggleEditorFileMaximized,
+    editorFilePath: activeEditorFilePath,
+    editorNavigationTarget,
+    editorHighlightTarget,
+    openEditorTabs: openFileTabs,
+    onActivateEditorTab: handleActivateWorkspaceFileTab,
+    onCloseEditorTab: handleCloseWorkspaceFileTab,
+    onCloseOtherEditorTabs: handleCloseOtherWorkspaceFileTabs,
+    onCloseAllEditorTabs: handleCloseAllWorkspaceFileTabs,
+    onReorderEditorTabs: handleReorderWorkspaceFileTabs,
+    onActiveEditorLineRangeChange: setActiveEditorLineRange,
+    onOpenFile: handleOpenWorkspaceFile,
+    onCompareFiles: handleOpenWorkspaceFileCompare,
+    onCloseFileCompare: handleCloseFileCompare,
+    externalChangeMonitoringEnabled: enableMainFileExternalChangeMonitoring,
+    externalChangeTransportMode: mainFileExternalChangeTransportMode,
+    externalChangeApplyMode: liveEditPreviewEnabled ? "auto" : "manual",
+    externalChangeAutoApplyDebounceMs: liveEditPreviewEnabled ? 700 : 0,
+    onExitEditor: handleExitWorkspaceEditor,
+    onExitDiff: handleExitDiff,
+    activeTab,
+    onSelectTab: setActiveTab,
+    tabletNavTab: tabletTab,
+    gitPanelMode,
+    onGitPanelModeChange: handleGitPanelModeChange,
+    onOpenGitHistoryPanel: handleOpenGitHistoryPanel,
+    onOpenProjectMap: handleOpenProjectMap,
+    gitDiffViewStyle,
+    gitDiffListView,
+    onGitDiffListViewChange: setGitDiffListView,
+    worktreeApplyLabel: t("git.applyWorktreeChangesAction"),
+    worktreeApplyTitle: activeParentWorkspace?.name
+      ? t("git.applyWorktreeChanges") + ` ${activeParentWorkspace.name}`
+      : t("git.applyWorktreeChanges"),
+    worktreeApplyLoading: isWorktreeWorkspace ? worktreeApplyLoading : false,
+    worktreeApplyError: isWorktreeWorkspace ? worktreeApplyError : null,
+    worktreeApplySuccess: isWorktreeWorkspace ? worktreeApplySuccess : false,
+    onApplyWorktreeChanges: isWorktreeWorkspace
+      ? handleApplyWorktreeChanges
+      : undefined,
+  };
+}
