+++ src/app-shell/sections/layoutNodes/layoutNodesOptionsGit.ts
@@ -1,0 +1,188 @@
+import type { GitLayoutNodesOptions } from "../../../features/layout/hooks/layoutNodesTypes";
+
+// B4 三期：useLayoutNodes 的 git options 对象逐字外移；键名/语句/注释保持原样，
+// hook 局部变量统一经 bag 传入（引用名不改）。
+export function buildLayoutNodesGitOptions(
+  bag: Record<string, any>,
+): GitLayoutNodesOptions {
+  const {
+    gitStatus,
+    fileStatus,
+    selectedDiffPath,
+    diffScrollRequestId,
+    handleSelectDiffForPanel,
+    gitLogEntries,
+    gitLogTotal,
+    gitLogAhead,
+    gitLogBehind,
+    gitLogAheadEntries,
+    gitLogBehindEntries,
+    gitLogUpstream,
+    gitLogError,
+    gitLogLoading,
+    selectedCommitSha,
+    gitIssues,
+    gitIssuesTotal,
+    gitIssuesLoading,
+    gitIssuesError,
+    gitPullRequests,
+    gitPullRequestsTotal,
+    gitPullRequestsLoading,
+    gitPullRequestsError,
+    selectedPullRequest,
+    diffSource,
+    gitPullRequestComments,
+    gitPullRequestCommentsLoading,
+    gitPullRequestCommentsError,
+    handleGitSelectPullRequest,
+    handleGitSelectCommit,
+    gitRemoteUrl,
+    activeGitRoot,
+    gitRootCandidates,
+    gitRootScanDepth,
+    gitRootScanLoading,
+    gitRootScanError,
+    gitRootScanHasScanned,
+    setGitRootScanDepth,
+    scanGitRoots,
+    handleSelectGitRoot,
+    handleClearGitRoot,
+    handlePickGitRoot,
+    handleStageGitAll,
+    handleStageGitFile,
+    handleUnstageGitAll,
+    handleUnstageGitFile,
+    handleUnstageGitPaths,
+    handleRevertGitFile,
+    handleRevertGitPaths,
+    handleRevertAllGitChanges,
+    activeDiffs,
+    activeDiffLoading,
+    activeDiffError,
+    refreshGitLog,
+    refreshGitDiffs,
+    handleManualGitStatusRefresh,
+    handleActiveDiffPath,
+    setGitDiffViewStyle,
+    commitMessage,
+    commitMessageLoading,
+    commitMessageError,
+    handleCommitMessageChange,
+    handleGenerateCommitMessage,
+    handleCommit,
+    handleCommitAndPush,
+    handleCommitAndSync,
+    handlePush,
+    handleSync,
+    commitLoading,
+    pushLoading,
+    syncLoading,
+    commitError,
+    pushError,
+    syncError,
+    isMultiRepository,
+    repositoryStatuses,
+    repositoryStatusesLoading,
+    refreshRepositoryStatuses,
+    handleStageRepositoryFile,
+    handleUnstageRepositoryFile,
+    handleUnstageRepositoryAll,
+    handleUnstageRepositoryFiles,
+    handleRevertRepositoryFile,
+    handleRevertRepositoryFiles,
+    handleStageRepositoryAll,
+    handleCommitRepositories,
+    repositoryCommitSummary,
+  } = bag;
+  return {
+    gitStatus,
+    fileStatus,
+    selectedDiffPath,
+    diffScrollRequestId,
+    onSelectDiff: handleSelectDiffForPanel,
+    gitLogEntries,
+    gitLogTotal,
+    gitLogAhead,
+    gitLogBehind,
+    gitLogAheadEntries,
+    gitLogBehindEntries,
+    gitLogUpstream,
+    gitLogError,
+    gitLogLoading,
+    selectedCommitSha,
+    gitIssues,
+    gitIssuesTotal,
+    gitIssuesLoading,
+    gitIssuesError,
+    gitPullRequests,
+    gitPullRequestsTotal,
+    gitPullRequestsLoading,
+    gitPullRequestsError,
+    selectedPullRequestNumber: selectedPullRequest?.number ?? null,
+    selectedPullRequest: diffSource === "pr" ? selectedPullRequest : null,
+    selectedPullRequestComments:
+      diffSource === "pr" ? gitPullRequestComments : [],
+    selectedPullRequestCommentsLoading: gitPullRequestCommentsLoading,
+    selectedPullRequestCommentsError: gitPullRequestCommentsError,
+    onSelectPullRequest: handleGitSelectPullRequest,
+    onSelectCommit: handleGitSelectCommit,
+    gitRemoteUrl,
+    gitRoot: activeGitRoot,
+    gitRootCandidates,
+    gitRootScanDepth,
+    gitRootScanLoading,
+    gitRootScanError,
+    gitRootScanHasScanned,
+    onGitRootScanDepthChange: setGitRootScanDepth,
+    onScanGitRoots: scanGitRoots,
+    onSelectGitRoot: handleSelectGitRoot,
+    onClearGitRoot: handleClearGitRoot,
+    onPickGitRoot: handlePickGitRoot,
+    onStageGitAll: handleStageGitAll,
+    onStageGitFile: handleStageGitFile,
+    onUnstageGitAll: handleUnstageGitAll,
+    onUnstageGitFile: handleUnstageGitFile,
+    onUnstageGitPaths: handleUnstageGitPaths,
+    onRevertGitFile: handleRevertGitFile,
+    onRevertGitPaths: handleRevertGitPaths,
+    onRevertAllGitChanges: handleRevertAllGitChanges,
+    gitDiffs: activeDiffs,
+    gitDiffLoading: activeDiffLoading,
+    gitDiffError: activeDiffError,
+    refreshGitLog,
+    refreshGitDiffs,
+    queueGitStatusRefresh: handleManualGitStatusRefresh,
+    onDiffActivePathChange: handleActiveDiffPath,
+    onGitDiffViewStyleChange: setGitDiffViewStyle,
+    commitMessage,
+    commitMessageLoading,
+    commitMessageError,
+    onCommitMessageChange: handleCommitMessageChange,
+    onGenerateCommitMessage: handleGenerateCommitMessage,
+    onCommit: handleCommit,
+    onCommitAndPush: handleCommitAndPush,
+    onCommitAndSync: handleCommitAndSync,
+    onPush: handlePush,
+    onSync: handleSync,
+    commitLoading,
+    pushLoading,
+    syncLoading,
+    commitError,
+    pushError,
+    syncError,
+    commitsAhead: gitLogAhead,
+    multiRepositoryMode: isMultiRepository,
+    repositoryStatuses,
+    repositoryStatusesLoading,
+    onRefreshRepositoryStatuses: refreshRepositoryStatuses,
+    onStageRepositoryFile: handleStageRepositoryFile,
+    onUnstageRepositoryFile: handleUnstageRepositoryFile,
+    onUnstageRepositoryAll: handleUnstageRepositoryAll,
+    onUnstageRepositoryFiles: handleUnstageRepositoryFiles,
+    onRevertRepositoryFile: handleRevertRepositoryFile,
+    onRevertRepositoryFiles: handleRevertRepositoryFiles,
+    onStageRepositoryAll: handleStageRepositoryAll,
+    onCommitRepositories: handleCommitRepositories,
+    repositoryCommitSummary,
+  };
+}
