+++ src/app-shell/sections/layoutNodes/layoutNodesOptionsPanels.ts
@@ -1,0 +1,112 @@
+import type { PanelsLayoutNodesOptions } from "../../../features/layout/hooks/layoutNodesTypes";
+
+// B4 三期：useLayoutNodes 的 panels options 对象逐字外移；键名/语句/注释保持原样，
+// hook 局部变量统一经 bag 传入（引用名不改）。
+export function buildLayoutNodesPanelsOptions(
+  bag: Record<string, any>,
+): PanelsLayoutNodesOptions {
+  const {
+    showComposer,
+    activePlan,
+    isPlanMode,
+    openPlanPanel,
+    closePlanPanel,
+    isPlanPanelDismissed,
+    agentTaskScrollRequest,
+    handleSelectStatusPanelSubagent,
+    debugEntries,
+    debugOpen,
+    terminalOpen,
+    terminalTabs,
+    activeTerminalId,
+    onSelectTerminal,
+    onNewTerminal,
+    onCloseTerminal,
+    terminalState,
+    clearDebugEntries,
+    handleCopyDebug,
+    onDebugPanelResizeStart,
+    onTerminalPanelResizeStart,
+    handleBackFromDiff,
+    handleGoProjects,
+    workspaceDropTargetRef,
+    dropOverlayActive,
+    dropOverlayText,
+    handleWorkspaceDragOver,
+    handleWorkspaceDragEnter,
+    handleWorkspaceDragLeave,
+    handleWorkspaceDrop,
+    appMode,
+    handleAppModeChange,
+    handleOpenHomeChat,
+    handleOpenMemory,
+    handleOpenProjectMemory,
+    handleOpenReleaseNotes,
+    focusedProjectMemoryId,
+    focusedProjectMemoryRequestKey,
+    focusedWorkspaceNoteId,
+    focusedWorkspaceNoteRequestKey,
+    handleOpenSearchPalette,
+    handleOpenQuickSwitcher,
+    collapseSidebar,
+    appSettings,
+    handleOpenWorkspaceHome,
+  } = bag;
+  return {
+    showComposer,
+    plan: activePlan,
+    isPlanMode,
+    onOpenPlanPanel: openPlanPanel,
+    onClosePlanPanel: closePlanPanel,
+    bottomStatusPanelExpanded: !isPlanPanelDismissed,
+    agentTaskScrollRequest,
+    onSelectSubagent: handleSelectStatusPanelSubagent,
+    debugEntries,
+    debugOpen,
+    terminalOpen,
+    terminalTabs,
+    activeTerminalId,
+    onSelectTerminal,
+    onNewTerminal,
+    onCloseTerminal,
+    terminalState,
+    onClearDebug: clearDebugEntries,
+    onCopyDebug: handleCopyDebug,
+    onResizeDebug: onDebugPanelResizeStart,
+    onResizeTerminal: onTerminalPanelResizeStart,
+    onBackFromDiff: handleBackFromDiff,
+    onGoProjects: handleGoProjects,
+    workspaceDropTargetRef,
+    isWorkspaceDropActive: dropOverlayActive,
+    workspaceDropText: dropOverlayText,
+    onWorkspaceDragOver: handleWorkspaceDragOver,
+    onWorkspaceDragEnter: handleWorkspaceDragEnter,
+    onWorkspaceDragLeave: handleWorkspaceDragLeave,
+    onWorkspaceDrop: handleWorkspaceDrop,
+    appMode,
+    onAppModeChange: handleAppModeChange,
+    onOpenHomeChat: handleOpenHomeChat,
+    onOpenMemory: handleOpenMemory,
+    onOpenProjectMemory: handleOpenProjectMemory,
+
+    onOpenReleaseNotes: handleOpenReleaseNotes,
+    focusedProjectMemoryId,
+    focusedProjectMemoryRequestKey,
+    focusedWorkspaceNoteId,
+    focusedWorkspaceNoteRequestKey,
+    onOpenGlobalSearch: handleOpenSearchPalette,
+    onOpenQuickSwitcher: handleOpenQuickSwitcher,
+    onCollapseSidebar: collapseSidebar,
+    globalSearchShortcut: appSettings.toggleGlobalSearchShortcut,
+    openChatShortcut: appSettings.openChatShortcut,
+    openSettingsShortcut: appSettings.openSettingsShortcut,
+    cycleOpenSessionPrevShortcut: appSettings.cycleOpenSessionPrevShortcut,
+    cycleOpenSessionNextShortcut: appSettings.cycleOpenSessionNextShortcut,
+    closeCurrentSessionShortcut: appSettings.closeCurrentSessionShortcut,
+    saveFileShortcut: appSettings.saveFileShortcut,
+    findInFileShortcut: appSettings.findInFileShortcut,
+    expandSelectionShortcut: appSettings.expandSelectionShortcut,
+    toggleGitDiffListViewShortcut: appSettings.toggleGitDiffListViewShortcut,
+    onOpenWorkspaceHome: handleOpenWorkspaceHome,
+  };
+}
