+++ src/app-shell/sections/layoutNodes/layoutNodesOptionsRuntime.ts
@@ -1,0 +1,71 @@
+import type { RuntimeLayoutNodesOptions } from "../../../features/layout/hooks/layoutNodesTypes";
+
+// B4 三期：useLayoutNodes 的 runtime options 对象逐字外移；键名/语句/注释保持原样，
+// hook 局部变量统一经 bag 传入（引用名不改）。
+export function buildLayoutNodesRuntimeOptions(
+  bag: Record<string, any>,
+): RuntimeLayoutNodesOptions {
+  const {
+    activeItems,
+    activeQueuedHandoffBubble,
+    threadItemsByThread,
+    sessionRadarRunningSessions,
+    sessionRadarRecentCompletedSessions,
+    activeRateLimits,
+    appSettings,
+    handleChangeDefaultVisibleThreadRootCount,
+    handleRefreshAccountRateLimits,
+    activeAccount,
+    handleSwitchAccount,
+    handleCancelSwitchAccount,
+    accountSwitching,
+    openAppIconById,
+    handleSelectOpenAppId,
+    approvals,
+    userInputRequests,
+    handleApprovalDecision,
+    handleApprovalBatchAccept,
+    handleApprovalRemember,
+    handleUserInputSubmitWithPlanApply,
+    handleUserInputDismiss,
+    handleRecoverThreadRuntime,
+    handleRecoverThreadRuntimeAndResend,
+    handleThreadRecoveryFork,
+    handleExitPlanModeExecute,
+  } = bag;
+  return {
+    activeItems,
+    activeQueuedHandoffBubble,
+    threadItemsByThread,
+    sessionRadarRunningSessions,
+    sessionRadarRecentCompletedSessions,
+    activeRateLimits,
+    usageShowRemaining: appSettings.usageShowRemaining,
+    showSidebarProviderLabels: appSettings.showSidebarProviderLabels,
+    defaultVisibleThreadRootCount: appSettings.defaultVisibleThreadRootCount,
+    onChangeDefaultVisibleThreadRootCount:
+      handleChangeDefaultVisibleThreadRootCount,
+    onRefreshAccountRateLimits: handleRefreshAccountRateLimits,
+    showMessageAnchors: appSettings.showMessageAnchors,
+    accountInfo: activeAccount,
+    onSwitchAccount: handleSwitchAccount,
+    onCancelSwitchAccount: handleCancelSwitchAccount,
+    accountSwitching,
+    codeBlockCopyUseModifier: appSettings.composerCodeBlockCopyUseModifier,
+    openAppTargets: appSettings.openAppTargets,
+    openAppIconById,
+    selectedOpenAppId: appSettings.selectedOpenAppId,
+    onSelectOpenAppId: handleSelectOpenAppId,
+    approvals,
+    userInputRequests,
+    handleApprovalDecision,
+    handleApprovalBatchAccept,
+    handleApprovalRemember,
+    handleUserInputSubmit: handleUserInputSubmitWithPlanApply,
+    handleUserInputDismiss,
+    onRecoverThreadRuntime: handleRecoverThreadRuntime,
+    onRecoverThreadRuntimeAndResend: handleRecoverThreadRuntimeAndResend,
+    onThreadRecoveryFork: handleThreadRecoveryFork,
+    handleExitPlanModeExecute,
+  };
+}
