+++ src/app-shell/sections/layoutNodes/useAppShellLayoutNodesSection.tsx
@@ -66,6 +66,13 @@ import {
 } from "../../../features/workspaces/utils/planWorkspaceNavigationThread";
 import { peekWorkspaceLastThreadId } from "../../../features/threads/utils/workspaceLastThreadMap";
 import { EMPTY_STRING_ARRAY, formatWorkspaceAliasError } from "./helpers";
+import { buildLayoutNodesChromeOptions } from "./layoutNodesOptionsChrome";
+import { buildLayoutNodesComposerOptions } from "./layoutNodesOptionsComposer";
+import { buildLayoutNodesEditorOptions } from "./layoutNodesOptionsEditor";
+import { buildLayoutNodesGitOptions } from "./layoutNodesOptionsGit";
+import { buildLayoutNodesPanelsOptions } from "./layoutNodesOptionsPanels";
+import { buildLayoutNodesRuntimeOptions } from "./layoutNodesOptionsRuntime";
+import { buildLayoutNodesWorkspaceOptions } from "./layoutNodesOptionsWorkspace";
 import {
   mergeAppShellDomainBag,
   selectAppShellDomainBag,
@@ -1964,533 +1971,177 @@ export function useAppShellLayoutNodesSection(
     compactGitBackNode,
     codeAnnotationBridgeProps,
   } = useLayoutNodes({
-    workspace: {
-      workspaces,
-      groupedWorkspaces,
-      hasWorkspaceGroups: workspaceGroups.length > 0,
-      deletingWorktreeIds,
-      threadsByWorkspace,
-      threadParentById,
-      threadStatusById,
-      historyLoadingByThreadId,
-      historyLoadingProgressByThreadId,
-      historyRestoredAtMsByThread,
-      runningSessionCountByWorkspaceId,
-      recentCompletedSessionCountByWorkspaceId,
-      // Prefer state snapshot (new Set identity on each mark) over the ref so
-      // memo(Sidebar) actually re-renders when hydration completes / times out.
-      hydratedThreadListWorkspaceIds:
-        hydratedThreadListWorkspaceIds instanceof Set
-          ? hydratedThreadListWorkspaceIds
-          : new Set<string>(),
-      threadListLoadingByWorkspace,
-      threadListPagingByWorkspace,
-      threadListCursorByWorkspace,
-      activeWorkspaceId,
-      activeThreadId,
-      isPhone,
-      isTablet,
-      systemProxyEnabled: appSettings.systemProxyEnabled,
-      systemProxyUrl: appSettings.systemProxyUrl,
-      onUpdateSystemProxy: handleUpdateSystemProxy,
-    },
-    runtime: {
-      activeItems,
-      activeQueuedHandoffBubble,
-      threadItemsByThread,
-      sessionRadarRunningSessions,
-      sessionRadarRecentCompletedSessions,
-      activeRateLimits,
-      usageShowRemaining: appSettings.usageShowRemaining,
-      showSidebarProviderLabels: appSettings.showSidebarProviderLabels,
-      defaultVisibleThreadRootCount: appSettings.defaultVisibleThreadRootCount,
-      onChangeDefaultVisibleThreadRootCount:
-        handleChangeDefaultVisibleThreadRootCount,
-      onRefreshAccountRateLimits: handleRefreshAccountRateLimits,
-      showMessageAnchors: appSettings.showMessageAnchors,
-      accountInfo: activeAccount,
-      onSwitchAccount: handleSwitchAccount,
-      onCancelSwitchAccount: handleCancelSwitchAccount,
-      accountSwitching,
-      codeBlockCopyUseModifier: appSettings.composerCodeBlockCopyUseModifier,
-      openAppTargets: appSettings.openAppTargets,
-      openAppIconById,
-      selectedOpenAppId: appSettings.selectedOpenAppId,
-      onSelectOpenAppId: handleSelectOpenAppId,
-      approvals,
-      userInputRequests,
-      handleApprovalDecision,
-      handleApprovalBatchAccept,
-      handleApprovalRemember,
-      handleUserInputSubmit: handleUserInputSubmitWithPlanApply,
-      handleUserInputDismiss,
-      onRecoverThreadRuntime: handleRecoverThreadRuntime,
-      onRecoverThreadRuntimeAndResend: handleRecoverThreadRuntimeAndResend,
-      onThreadRecoveryFork: handleThreadRecoveryFork,
-      handleExitPlanModeExecute,
-    },
-    chrome: {
-      onOpenSettings: handleOpenSettings,
-      onOpenSessionManagement: handleOpenSessionManagement,
-      onOpenShortcutsSettings: handleOpenShortcutsSettings,
-      onOpenAgentSettings: handleOpenAgentSettings,
-      onOpenPromptSettings: handleOpenPromptSettings,
-      onOpenModelSettings: handleOpenModelSettings,
-      onOpenCliSettings: handleOpenCliSettings,
-      onRefreshModelConfig: handleRefreshModelConfig,
-      isModelConfigRefreshing,
-      onOpenSkillsSettings: handleOpenSkillsSettings,
-      onOpenDebug: handleDebugClick,
-      showDebugButton,
-      onAddWorkspace: handleAddWorkspace,
-      onSelectHome: handleSelectHome,
-      onSelectWorkspace: handleSelectWorkspace,
-      onReorderWorkspaces: handleReorderWorkspaces,
-      onConnectWorkspace: handleConnectWorkspace,
-      onAddAgent: handleAddAgent,
-      engineOptions: availableEngines,
-      onRefreshEngineOptions: refreshEngines,
-      onAddSharedAgent: handleStartSharedConversation,
-      onAddWorktreeAgent: handleAddWorktreeAgent,
-      onAddCloneAgent: handleAddCloneAgent,
-      onToggleWorkspaceCollapse: handleToggleWorkspaceCollapse,
-      onSelectThread: handleSelectThread,
-      onClearActiveThread: handleClearActiveThread,
-      onProviderContinuationTargetReady: handleProviderContinuationTargetReady,
-      onSelectHomeWorkspace: handleSelectHomeWorkspace,
-      onDeleteThread: handleDeleteThread,
-      onArchiveThread: handleArchiveThread,
-      deleteConfirmThreadId: deleteThreadPrompt?.threadId ?? null,
-      deleteConfirmWorkspaceId: deleteThreadPrompt?.workspaceId ?? null,
-      deleteConfirmBusy: isDeleteThreadPromptBusy,
-      onCancelDeleteConfirm: handleDeleteThreadPromptCancel,
-      onConfirmDeleteConfirm: handleConfirmDeleteConfirm,
-      renameThreadId: renamePrompt?.threadId ?? null,
-      renameWorkspaceId: renamePrompt?.workspaceId ?? null,
-      renameName: renamePrompt?.name ?? "",
-      onRenameChange: handleRenamePromptChange,
-      onRenameCancel: handleRenamePromptCancel,
-      onRenameConfirm: handleRenamePromptConfirm,
-      onSyncThread: handleSyncThread,
-      pinThread,
-      unpinThread,
-      isThreadPinned,
-      getPinTimestamp,
-      pinnedThreadsVersion,
-      isThreadAutoNaming,
-      onRenameThread: handleSidebarRenameThread,
-      onAutoNameThread: handleAutoNameThread,
-      onOpenClaudeTui: handleOpenClaudeTui,
-      onDeleteWorkspace: handleDeleteWorkspace,
-      onDeleteWorktree: handleDeleteWorktree,
-      onRenameWorkspaceAlias: handleRenameWorkspaceAlias,
-      workspaceGroups,
-      onAssignWorkspaceGroup: assignWorkspaceGroup,
-      onLoadOlderThreads: handleLoadOlderThreads,
-      onQuickReloadWorkspaceThreads: handleQuickReloadWorkspaceThreads,
-      onReloadWorkspaceThreads: handleReloadWorkspaceThreads,
-      isExitedSessionsHidden,
-      onToggleExitedSessionsHidden: toggleExitedSessionsHidden,
+    workspace: buildLayoutNodesWorkspaceOptions({
+      workspaces, groupedWorkspaces, workspaceGroups, deletingWorktreeIds,
+      threadsByWorkspace, threadParentById, threadStatusById,
+      historyLoadingByThreadId, historyLoadingProgressByThreadId,
+      historyRestoredAtMsByThread, runningSessionCountByWorkspaceId,
+      recentCompletedSessionCountByWorkspaceId, hydratedThreadListWorkspaceIds,
+      threadListLoadingByWorkspace, threadListPagingByWorkspace,
+      threadListCursorByWorkspace, activeWorkspaceId, activeThreadId, isPhone,
+      isTablet, appSettings, handleUpdateSystemProxy,
+    }),
+    runtime: buildLayoutNodesRuntimeOptions({
+      activeItems, activeQueuedHandoffBubble, threadItemsByThread,
+      sessionRadarRunningSessions, sessionRadarRecentCompletedSessions,
+      activeRateLimits, appSettings, handleChangeDefaultVisibleThreadRootCount,
+      handleRefreshAccountRateLimits, activeAccount, handleSwitchAccount,
+      handleCancelSwitchAccount, accountSwitching, openAppIconById,
+      handleSelectOpenAppId, approvals, userInputRequests,
+      handleApprovalDecision, handleApprovalBatchAccept, handleApprovalRemember,
+      handleUserInputSubmitWithPlanApply, handleUserInputDismiss,
+      handleRecoverThreadRuntime, handleRecoverThreadRuntimeAndResend,
+      handleThreadRecoveryFork, handleExitPlanModeExecute,
+    }),
+    chrome: buildLayoutNodesChromeOptions({
+      handleOpenSettings, handleOpenSessionManagement,
+      handleOpenShortcutsSettings, handleOpenAgentSettings,
+      handleOpenPromptSettings, handleOpenModelSettings, handleOpenCliSettings,
+      handleRefreshModelConfig, isModelConfigRefreshing,
+      handleOpenSkillsSettings, handleDebugClick, showDebugButton,
+      handleAddWorkspace, handleSelectHome, handleSelectWorkspace,
+      handleReorderWorkspaces, handleConnectWorkspace, handleAddAgent,
+      availableEngines, refreshEngines, handleStartSharedConversation,
+      handleAddWorktreeAgent, handleAddCloneAgent,
+      handleToggleWorkspaceCollapse, handleSelectThread,
+      handleClearActiveThread, handleProviderContinuationTargetReady,
+      handleSelectHomeWorkspace, handleDeleteThread, handleArchiveThread,
+      deleteThreadPrompt, isDeleteThreadPromptBusy,
+      handleDeleteThreadPromptCancel, handleConfirmDeleteConfirm, renamePrompt,
+      handleRenamePromptChange, handleRenamePromptCancel,
+      handleRenamePromptConfirm, handleSyncThread, pinThread, unpinThread,
+      isThreadPinned, getPinTimestamp, pinnedThreadsVersion, isThreadAutoNaming,
+      handleSidebarRenameThread, handleAutoNameThread, handleOpenClaudeTui,
+      handleDeleteWorkspace, handleDeleteWorktree, handleRenameWorkspaceAlias,
+      workspaceGroups, assignWorkspaceGroup, handleLoadOlderThreads,
+      handleQuickReloadWorkspaceThreads, handleReloadWorkspaceThreads,
+      isExitedSessionsHidden, toggleExitedSessionsHidden,
       rootSessionFolderDraftRequestByWorkspaceId,
-      onRequestRootSessionFolderDraft,
-      updaterState,
-      onUpdate: startUpdate,
-      onDismissUpdate: dismissUpdate,
-      errorToasts,
-      onDismissErrorToast: dismissErrorToast,
-      onOpenSpecHub: handleOpenSpecHub,
-      showLoadingProgressDialog,
-      hideLoadingProgressDialog,
-      activeWorkspace,
-      activeParentWorkspace,
-      worktreeLabel,
-      worktreeRename: worktreeRename ?? undefined,
-      isWorktreeWorkspace,
-      branchName: gitStatus.branchName,
-      branches,
-      branchError,
-      branchCurrentName: currentBranch,
-      branchLocalItems: localBranches,
-      branchRemoteItems: remoteBranches,
-      gitRepositories: repositories,
-      gitRepositoriesLoading: repositoriesLoading,
-      gitRepositoriesError: repositoryError,
-      selectedGitRepositoryRoot: selectedRepositoryRoot,
-      onSelectGitRepository: selectRepository,
-      onCheckoutBranch: handleCheckoutBranch,
-      onCreateBranch: handleCreateBranch,
-      onUpdateBranch: handleUpdateBranch,
-      onUpdateAllRepositories: handleUpdateAllRepositories,
-      onCheckoutAllRepositories: handleCheckoutAllRepositories,
-      onLoadCommonRepositoryBranches: handleLoadCommonRepositoryBranches,
-      onCopyThread: handleCopyThread,
-      onLockPanel: handleLockPanel,
-      onToggleTerminal: handleToggleTerminalPanel,
-      showTerminalButton: !isCompact,
-      launchScript: launchScriptState.launchScript,
-      launchScriptEditorOpen: launchScriptState.editorOpen,
-      launchScriptDraft: launchScriptState.draftScript,
-      launchScriptSaving: launchScriptState.isSaving,
-      launchScriptError: launchScriptState.error,
-      onRunLaunchScript: launchScriptState.onRunLaunchScript,
-      onOpenLaunchScriptEditor: launchScriptState.onOpenEditor,
-      onCloseLaunchScriptEditor: launchScriptState.onCloseEditor,
-      onLaunchScriptDraftChange: launchScriptState.onDraftScriptChange,
-      onSaveLaunchScript: launchScriptState.onSaveLaunchScript,
-      launchScriptsState,
-      mainHeaderActions,
-      filePanelMode,
-      onFilePanelModeChange: setFilePanelMode,
-      liveEditPreviewEnabled,
-      onToggleLiveEditPreview: handleToggleLiveEditPreview,
-      fileTreeLoading: isFilesLoading,
-      fileTreeLoadError,
-      onRefreshFiles: refreshFiles,
-      onOpenDetachedFileExplorer: handleOpenDetachedFileExplorer,
-      onToggleRuntimeConsole: handleToggleRuntimeConsole,
-      runtimeConsoleVisible: runtimeRunState.runtimeConsoleVisible,
-      browserDockOpen,
-      onCloseBrowserDock: handleCloseBrowserDock,
-    },
-    editor: {
-      centerMode,
-      setCenterMode,
-      fileCompareSession,
-      onOpenFileHistory: handleOpenFileHistory,
-      editorSplitCompanion,
-      setEditorSplitCompanion,
-      editorSplitLayout,
-      onToggleEditorSplitLayout: handleToggleEditorSplitLayout,
-      isEditorFileMaximized,
-      onToggleEditorFileMaximized: handleToggleEditorFileMaximized,
-      editorFilePath: activeEditorFilePath,
-      editorNavigationTarget,
-      editorHighlightTarget,
-      openEditorTabs: openFileTabs,
-      onActivateEditorTab: handleActivateWorkspaceFileTab,
-      onCloseEditorTab: handleCloseWorkspaceFileTab,
-      onCloseOtherEditorTabs: handleCloseOtherWorkspaceFileTabs,
-      onCloseAllEditorTabs: handleCloseAllWorkspaceFileTabs,
-      onReorderEditorTabs: handleReorderWorkspaceFileTabs,
-      onActiveEditorLineRangeChange: setActiveEditorLineRange,
-      onOpenFile: handleOpenWorkspaceFile,
-      onCompareFiles: handleOpenWorkspaceFileCompare,
-      onCloseFileCompare: handleCloseFileCompare,
-      externalChangeMonitoringEnabled: enableMainFileExternalChangeMonitoring,
-      externalChangeTransportMode: mainFileExternalChangeTransportMode,
-      externalChangeApplyMode: liveEditPreviewEnabled ? "auto" : "manual",
-      externalChangeAutoApplyDebounceMs: liveEditPreviewEnabled ? 700 : 0,
-      onExitEditor: handleExitWorkspaceEditor,
-      onExitDiff: handleExitDiff,
-      activeTab,
-      onSelectTab: setActiveTab,
-      tabletNavTab: tabletTab,
-      gitPanelMode,
-      onGitPanelModeChange: handleGitPanelModeChange,
-      onOpenGitHistoryPanel: handleOpenGitHistoryPanel,
-      onOpenProjectMap: handleOpenProjectMap,
-      gitDiffViewStyle,
-      gitDiffListView,
-      onGitDiffListViewChange: setGitDiffListView,
-      worktreeApplyLabel: t("git.applyWorktreeChangesAction"),
-      worktreeApplyTitle: activeParentWorkspace?.name
-        ? t("git.applyWorktreeChanges") + ` ${activeParentWorkspace.name}`
-        : t("git.applyWorktreeChanges"),
-      worktreeApplyLoading: isWorktreeWorkspace ? worktreeApplyLoading : false,
-      worktreeApplyError: isWorktreeWorkspace ? worktreeApplyError : null,
-      worktreeApplySuccess: isWorktreeWorkspace ? worktreeApplySuccess : false,
-      onApplyWorktreeChanges: isWorktreeWorkspace
-        ? handleApplyWorktreeChanges
-        : undefined,
-    },
-    git: {
-      gitStatus,
-      fileStatus,
-      selectedDiffPath,
-      diffScrollRequestId,
-      onSelectDiff: handleSelectDiffForPanel,
-      gitLogEntries,
-      gitLogTotal,
-      gitLogAhead,
-      gitLogBehind,
-      gitLogAheadEntries,
-      gitLogBehindEntries,
-      gitLogUpstream,
-      gitLogError,
-      gitLogLoading,
-      selectedCommitSha,
-      gitIssues,
-      gitIssuesTotal,
-      gitIssuesLoading,
-      gitIssuesError,
-      gitPullRequests,
-      gitPullRequestsTotal,
-      gitPullRequestsLoading,
-      gitPullRequestsError,
-      selectedPullRequestNumber: selectedPullRequest?.number ?? null,
-      selectedPullRequest: diffSource === "pr" ? selectedPullRequest : null,
-      selectedPullRequestComments:
-        diffSource === "pr" ? gitPullRequestComments : [],
-      selectedPullRequestCommentsLoading: gitPullRequestCommentsLoading,
-      selectedPullRequestCommentsError: gitPullRequestCommentsError,
-      onSelectPullRequest: handleGitSelectPullRequest,
-      onSelectCommit: handleGitSelectCommit,
-      gitRemoteUrl,
-      gitRoot: activeGitRoot,
-      gitRootCandidates,
-      gitRootScanDepth,
-      gitRootScanLoading,
-      gitRootScanError,
-      gitRootScanHasScanned,
-      onGitRootScanDepthChange: setGitRootScanDepth,
-      onScanGitRoots: scanGitRoots,
-      onSelectGitRoot: handleSelectGitRoot,
-      onClearGitRoot: handleClearGitRoot,
-      onPickGitRoot: handlePickGitRoot,
-      onStageGitAll: handleStageGitAll,
-      onStageGitFile: handleStageGitFile,
-      onUnstageGitAll: handleUnstageGitAll,
-      onUnstageGitFile: handleUnstageGitFile,
-      onUnstageGitPaths: handleUnstageGitPaths,
-      onRevertGitFile: handleRevertGitFile,
-      onRevertGitPaths: handleRevertGitPaths,
-      onRevertAllGitChanges: handleRevertAllGitChanges,
-      gitDiffs: activeDiffs,
-      gitDiffLoading: activeDiffLoading,
-      gitDiffError: activeDiffError,
-      refreshGitLog,
-      refreshGitDiffs,
-      queueGitStatusRefresh: handleManualGitStatusRefresh,
-      onDiffActivePathChange: handleActiveDiffPath,
-      onGitDiffViewStyleChange: setGitDiffViewStyle,
-      commitMessage,
-      commitMessageLoading,
-      commitMessageError,
-      onCommitMessageChange: handleCommitMessageChange,
-      onGenerateCommitMessage: handleGenerateCommitMessage,
-      onCommit: handleCommit,
-      onCommitAndPush: handleCommitAndPush,
-      onCommitAndSync: handleCommitAndSync,
-      onPush: handlePush,
-      onSync: handleSync,
-      commitLoading,
-      pushLoading,
-      syncLoading,
-      commitError,
-      pushError,
-      syncError,
-      commitsAhead: gitLogAhead,
-      multiRepositoryMode: isMultiRepository,
-      repositoryStatuses,
-      repositoryStatusesLoading,
-      onRefreshRepositoryStatuses: refreshRepositoryStatuses,
-      onStageRepositoryFile: handleStageRepositoryFile,
-      onUnstageRepositoryFile: handleUnstageRepositoryFile,
-      onUnstageRepositoryAll: handleUnstageRepositoryAll,
-      onUnstageRepositoryFiles: handleUnstageRepositoryFiles,
-      onRevertRepositoryFile: handleRevertRepositoryFile,
-      onRevertRepositoryFiles: handleRevertRepositoryFiles,
-      onStageRepositoryAll: handleStageRepositoryAll,
-      onCommitRepositories: handleCommitRepositories,
+      onRequestRootSessionFolderDraft, updaterState, startUpdate, dismissUpdate,
+      errorToasts, dismissErrorToast, handleOpenSpecHub,
+      showLoadingProgressDialog, hideLoadingProgressDialog, activeWorkspace,
+      activeParentWorkspace, worktreeLabel, worktreeRename, isWorktreeWorkspace,
+      gitStatus, branches, branchError, currentBranch, localBranches,
+      remoteBranches, repositories, repositoriesLoading, repositoryError,
+      selectedRepositoryRoot, selectRepository, handleCheckoutBranch,
+      handleCreateBranch, handleUpdateBranch, handleUpdateAllRepositories,
+      handleCheckoutAllRepositories, handleLoadCommonRepositoryBranches,
+      handleCopyThread, handleLockPanel, handleToggleTerminalPanel, isCompact,
+      launchScriptState, launchScriptsState, mainHeaderActions, filePanelMode,
+      setFilePanelMode, liveEditPreviewEnabled, handleToggleLiveEditPreview,
+      isFilesLoading, fileTreeLoadError, refreshFiles,
+      handleOpenDetachedFileExplorer, handleToggleRuntimeConsole,
+      runtimeRunState, browserDockOpen, handleCloseBrowserDock,
+    }),
+    editor: buildLayoutNodesEditorOptions({
+      centerMode, setCenterMode, fileCompareSession, handleOpenFileHistory,
+      editorSplitCompanion, setEditorSplitCompanion, editorSplitLayout,
+      handleToggleEditorSplitLayout, isEditorFileMaximized,
+      handleToggleEditorFileMaximized, activeEditorFilePath,
+      editorNavigationTarget, editorHighlightTarget, openFileTabs,
+      handleActivateWorkspaceFileTab, handleCloseWorkspaceFileTab,
+      handleCloseOtherWorkspaceFileTabs, handleCloseAllWorkspaceFileTabs,
+      handleReorderWorkspaceFileTabs, setActiveEditorLineRange,
+      handleOpenWorkspaceFile, handleOpenWorkspaceFileCompare,
+      handleCloseFileCompare, enableMainFileExternalChangeMonitoring,
+      mainFileExternalChangeTransportMode, liveEditPreviewEnabled,
+      handleExitWorkspaceEditor, handleExitDiff, activeTab, setActiveTab,
+      tabletTab, gitPanelMode, handleGitPanelModeChange,
+      handleOpenGitHistoryPanel, handleOpenProjectMap, gitDiffViewStyle,
+      gitDiffListView, setGitDiffListView, t, activeParentWorkspace,
+      isWorktreeWorkspace, worktreeApplyLoading, worktreeApplyError,
+      worktreeApplySuccess, handleApplyWorktreeChanges,
+    }),
+    git: buildLayoutNodesGitOptions({
+      gitStatus, fileStatus, selectedDiffPath, diffScrollRequestId,
+      handleSelectDiffForPanel, gitLogEntries, gitLogTotal, gitLogAhead,
+      gitLogBehind, gitLogAheadEntries, gitLogBehindEntries, gitLogUpstream,
+      gitLogError, gitLogLoading, selectedCommitSha, gitIssues, gitIssuesTotal,
+      gitIssuesLoading, gitIssuesError, gitPullRequests, gitPullRequestsTotal,
+      gitPullRequestsLoading, gitPullRequestsError, selectedPullRequest,
+      diffSource, gitPullRequestComments, gitPullRequestCommentsLoading,
+      gitPullRequestCommentsError, handleGitSelectPullRequest,
+      handleGitSelectCommit, gitRemoteUrl, activeGitRoot, gitRootCandidates,
+      gitRootScanDepth, gitRootScanLoading, gitRootScanError,
+      gitRootScanHasScanned, setGitRootScanDepth, scanGitRoots,
+      handleSelectGitRoot, handleClearGitRoot, handlePickGitRoot,
+      handleStageGitAll, handleStageGitFile, handleUnstageGitAll,
+      handleUnstageGitFile, handleUnstageGitPaths, handleRevertGitFile,
+      handleRevertGitPaths, handleRevertAllGitChanges, activeDiffs,
+      activeDiffLoading, activeDiffError, refreshGitLog, refreshGitDiffs,
+      handleManualGitStatusRefresh, handleActiveDiffPath, setGitDiffViewStyle,
+      commitMessage, commitMessageLoading, commitMessageError,
+      handleCommitMessageChange, handleGenerateCommitMessage, handleCommit,
+      handleCommitAndPush, handleCommitAndSync, handlePush, handleSync,
+      commitLoading, pushLoading, syncLoading, commitError, pushError,
+      syncError, isMultiRepository, repositoryStatuses,
+      repositoryStatusesLoading, refreshRepositoryStatuses,
+      handleStageRepositoryFile, handleUnstageRepositoryFile,
+      handleUnstageRepositoryAll, handleUnstageRepositoryFiles,
+      handleRevertRepositoryFile, handleRevertRepositoryFiles,
+      handleStageRepositoryAll, handleCommitRepositories,
       repositoryCommitSummary,
-    },
-    composer: {
-      onSendPrompt: handleSendPrompt,
-      onSendPromptToNewAgent: handleSendPromptToNewAgent,
-      onCreatePrompt: handleCreatePrompt,
-      onUpdatePrompt: handleUpdatePrompt,
-      onDeletePrompt: handleDeletePrompt,
-      onMovePrompt: handleMovePrompt,
-      onRevealWorkspacePrompts: handleRevealWorkspacePrompts,
-      onRevealGeneralPrompts: handleRevealGeneralPrompts,
-      canRevealGeneralPrompts: Boolean(activeWorkspace),
-      onSend: handleComposerSendWithIntentCanvas,
-      onQueue: handleComposerQueueWithIntentCanvas,
-      onRequestContextCompaction: handleRequestContextCompaction,
-      onStop: interruptTurn,
-      // Shared CLI：完成邮件与 Native turn lifecycle 脱节，隐藏入口避免假能力。
-      // ContextBar 以 onToggleCompletionEmail 是否传入决定是否渲染邮件图标。
-      completionEmailSelected: Boolean(
-        activeThreadId &&
-          !isSharedSessionThreadId(activeThreadId) &&
-          completionEmailIntentByThread?.[activeThreadId],
-      ),
-      completionEmailDisabled:
-        !activeThreadId || isSharedSessionThreadId(activeThreadId),
-      onToggleCompletionEmail: isSharedSessionThreadId(activeThreadId)
-        ? undefined
-        : handleToggleCompletionEmail,
-      onRewind: handleRewindFromMessage,
-      onForkFromMessage: handleForkFromMessage,
-      canStop: canInterrupt,
-      isReviewing,
-      isProcessing,
-      steerEnabled: appSettings.experimentalSteerEnabled,
-      reviewPrompt,
-      onReviewPromptClose: closeReviewPrompt,
-      onReviewPromptShowPreset: showPresetStep,
-      onReviewPromptChoosePreset: choosePreset,
-      highlightedPresetIndex,
-      onReviewPromptHighlightPreset: setHighlightedPresetIndex,
-      highlightedBranchIndex,
-      onReviewPromptHighlightBranch: setHighlightedBranchIndex,
-      highlightedCommitIndex,
-      onReviewPromptHighlightCommit: setHighlightedCommitIndex,
-      onReviewPromptKeyDown: handleReviewPromptKeyDown,
-      onReviewPromptSelectBranch: selectBranch,
-      onReviewPromptSelectBranchAtIndex: selectBranchAtIndex,
-      onReviewPromptConfirmBranch: confirmBranch,
-      onReviewPromptSelectCommit: selectCommit,
-      onReviewPromptSelectCommitAtIndex: selectCommitAtIndex,
-      onReviewPromptConfirmCommit: confirmCommit,
-      onReviewPromptUpdateCustomInstructions: updateCustomInstructions,
-      onReviewPromptConfirmCustom: confirmCustom,
-      activeTokenUsage,
-      contextDualViewEnabled: activeEngine === "codex",
-      codexAutoCompactionEnabled: appSettings.codexAutoCompactionEnabled,
-      codexAutoCompactionThresholdPercent:
-        appSettings.codexAutoCompactionThresholdPercent,
-      onCodexAutoCompactionSettingsChange:
-        handleCodexAutoCompactionSettingsChange,
-      activeQueue,
-      onDraftChange: handleDraftChange,
-      activeImages,
-      onPickImages: pickImages,
-      onAttachImages: attachImages,
-      onRemoveImage: removeImage,
-      prefillDraft,
-      onPrefillHandled: handlePrefillHandled,
-      insertText: composerInsert,
-      onInsertHandled: handleInsertHandled,
-      onEditQueued: handleEditQueued,
-      onDeleteQueued: handleDeleteQueued,
-      onFuseQueued: handleFuseQueued,
-      canFuseActiveQueue,
-      fuseDisabledReasonKey,
-      activeFusingMessageId,
-      collaborationModes,
-      collaborationModesEnabled,
-      selectedCollaborationModeId,
-      onSelectCollaborationMode: applySelectedCollaborationMode,
-      engines: availableEngines,
-      selectedEngine: activeEngine,
-      usePresentationProfile: appSettings.chatCanvasUsePresentationProfile,
-      onSelectEngine: handleSelectConversationEngine,
-      models: effectiveModels,
-      providerModelCatalogs,
-      selectedModelId: effectiveSelectedModelId,
-      projectMapDatasetController,
-      onSelectModel: handleSelectModel,
-      intentCanvasOpenRequest,
-      onOpenIntentCanvas: handleOpenIntentCanvas,
-      onIntentCanvasOpenRequestConsumed: handleIntentCanvasOpenRequestConsumed,
-      onAttachIntentCanvasToThread: handleAttachIntentCanvasToThread,
-      pendingIntentCanvasDocuments,
-      onRemovePendingIntentCanvas: handleRemovePendingIntentCanvas,
-      reasoningOptions,
-      selectedEffort,
-      onSelectEffort: setSelectedEffort,
-      claudeThinkingVisible,
-      onResolvedClaudeThinkingVisibleChange:
-        handleResolvedClaudeThinkingVisibleChange,
-      reasoningSupported: effectiveReasoningSupported,
-      opencodeAgents: openCodeAgents,
-      selectedOpenCodeAgent,
-      onSelectOpenCodeAgent: handleSelectOpenCodeAgent,
-      selectedAgent,
-      onSelectAgent: handleSelectAgent,
-      opencodeVariantOptions: OPENCODE_VARIANT_OPTIONS,
-      selectedOpenCodeVariant,
-      onSelectOpenCodeVariant: handleSelectOpenCodeVariant,
-      accessMode,
-      onSelectAccessMode: handleSetAccessMode,
-      skills,
-      customSkillDirectories:
-        appSettings.customSkillDirectories ?? EMPTY_STRING_ARRAY,
-      prompts,
-      commands,
-      files,
-      directories,
-      directoryMetadata,
-      fileTreeSourceVersion,
-      gitignoredFiles,
-      gitignoredDirectories,
-      onInsertComposerText: handleInsertComposerText,
-      textareaRef: composerInputRef,
-      composerEditorSettings,
-      composerSendShortcut: appSettings.composerSendShortcut,
-      composerInterruptShortcutLabel: formatShortcutLabelOrNull(
-        appSettings.interruptShortcut,
-      ),
-      textareaHeight,
-      onTextareaHeightChange,
-      onOpenExperimentalSettings: handleOpenExperimentalSettings,
-      composerSendLabel,
-      activeComposerFilePath: activeEditorFilePath,
-      activeComposerFileLineRange: activeEditorLineRange,
-      activeCodeSelectionAnchor: activeIntentCanvasCodeSelectionAnchor,
-      onActiveCodeSelectionAnchorChange:
-        setActiveIntentCanvasCodeSelectionAnchor,
-      fileReferenceMode,
-      onFileReferenceModeChange: setFileReferenceMode,
-    },
-    panels: {
-      showComposer,
-      plan: activePlan,
-      isPlanMode,
-      onOpenPlanPanel: openPlanPanel,
-      onClosePlanPanel: closePlanPanel,
-      bottomStatusPanelExpanded: !isPlanPanelDismissed,
-      agentTaskScrollRequest,
-      onSelectSubagent: handleSelectStatusPanelSubagent,
-      debugEntries,
-      debugOpen,
-      terminalOpen,
-      terminalTabs,
-      activeTerminalId,
-      onSelectTerminal,
-      onNewTerminal,
-      onCloseTerminal,
-      terminalState,
-      onClearDebug: clearDebugEntries,
-      onCopyDebug: handleCopyDebug,
-      onResizeDebug: onDebugPanelResizeStart,
-      onResizeTerminal: onTerminalPanelResizeStart,
-      onBackFromDiff: handleBackFromDiff,
-      onGoProjects: handleGoProjects,
-      workspaceDropTargetRef,
-      isWorkspaceDropActive: dropOverlayActive,
-      workspaceDropText: dropOverlayText,
-      onWorkspaceDragOver: handleWorkspaceDragOver,
-      onWorkspaceDragEnter: handleWorkspaceDragEnter,
-      onWorkspaceDragLeave: handleWorkspaceDragLeave,
-      onWorkspaceDrop: handleWorkspaceDrop,
-      appMode,
-      onAppModeChange: handleAppModeChange,
-      onOpenHomeChat: handleOpenHomeChat,
-      onOpenMemory: handleOpenMemory,
-      onOpenProjectMemory: handleOpenProjectMemory,
-
-      onOpenReleaseNotes: handleOpenReleaseNotes,
-      focusedProjectMemoryId,
-      focusedProjectMemoryRequestKey,
-      focusedWorkspaceNoteId,
-      focusedWorkspaceNoteRequestKey,
-      onOpenGlobalSearch: handleOpenSearchPalette,
-      onOpenQuickSwitcher: handleOpenQuickSwitcher,
-      onCollapseSidebar: collapseSidebar,
-      globalSearchShortcut: appSettings.toggleGlobalSearchShortcut,
-      openChatShortcut: appSettings.openChatShortcut,
-      openSettingsShortcut: appSettings.openSettingsShortcut,
-      cycleOpenSessionPrevShortcut: appSettings.cycleOpenSessionPrevShortcut,
-      cycleOpenSessionNextShortcut: appSettings.cycleOpenSessionNextShortcut,
-      closeCurrentSessionShortcut: appSettings.closeCurrentSessionShortcut,
-      saveFileShortcut: appSettings.saveFileShortcut,
-      findInFileShortcut: appSettings.findInFileShortcut,
-      expandSelectionShortcut: appSettings.expandSelectionShortcut,
-      toggleGitDiffListViewShortcut: appSettings.toggleGitDiffListViewShortcut,
-      onOpenWorkspaceHome: handleOpenWorkspaceHome,
-    },
+    }),
+    composer: buildLayoutNodesComposerOptions({
+      handleSendPrompt, handleSendPromptToNewAgent, handleCreatePrompt,
+      handleUpdatePrompt, handleDeletePrompt, handleMovePrompt,
+      handleRevealWorkspacePrompts, handleRevealGeneralPrompts, activeWorkspace,
+      handleComposerSendWithIntentCanvas, handleComposerQueueWithIntentCanvas,
+      handleRequestContextCompaction, interruptTurn, activeThreadId,
+      isSharedSessionThreadId, completionEmailIntentByThread,
+      handleToggleCompletionEmail, handleRewindFromMessage,
+      handleForkFromMessage, canInterrupt, isReviewing, isProcessing,
+      appSettings, reviewPrompt, closeReviewPrompt, showPresetStep,
+      choosePreset, highlightedPresetIndex, setHighlightedPresetIndex,
+      highlightedBranchIndex, setHighlightedBranchIndex, highlightedCommitIndex,
+      setHighlightedCommitIndex, handleReviewPromptKeyDown, selectBranch,
+      selectBranchAtIndex, confirmBranch, selectCommit, selectCommitAtIndex,
+      confirmCommit, updateCustomInstructions, confirmCustom, activeTokenUsage,
+      activeEngine, handleCodexAutoCompactionSettingsChange, activeQueue,
+      handleDraftChange, activeImages, pickImages, attachImages, removeImage,
+      prefillDraft, handlePrefillHandled, composerInsert, handleInsertHandled,
+      handleEditQueued, handleDeleteQueued, handleFuseQueued,
+      canFuseActiveQueue, fuseDisabledReasonKey, activeFusingMessageId,
+      collaborationModes, collaborationModesEnabled,
+      selectedCollaborationModeId, applySelectedCollaborationMode,
+      availableEngines, handleSelectConversationEngine, effectiveModels,
+      providerModelCatalogs, effectiveSelectedModelId,
+      projectMapDatasetController, handleSelectModel, intentCanvasOpenRequest,
+      handleOpenIntentCanvas, handleIntentCanvasOpenRequestConsumed,
+      handleAttachIntentCanvasToThread, pendingIntentCanvasDocuments,
+      handleRemovePendingIntentCanvas, reasoningOptions, selectedEffort,
+      setSelectedEffort, claudeThinkingVisible,
+      handleResolvedClaudeThinkingVisibleChange, effectiveReasoningSupported,
+      openCodeAgents, selectedOpenCodeAgent, handleSelectOpenCodeAgent,
+      selectedAgent, handleSelectAgent, OPENCODE_VARIANT_OPTIONS,
+      selectedOpenCodeVariant, handleSelectOpenCodeVariant, accessMode,
+      handleSetAccessMode, skills, EMPTY_STRING_ARRAY, prompts, commands, files,
+      directories, directoryMetadata, fileTreeSourceVersion, gitignoredFiles,
+      gitignoredDirectories, handleInsertComposerText, composerInputRef,
+      composerEditorSettings, formatShortcutLabelOrNull, textareaHeight,
+      onTextareaHeightChange, handleOpenExperimentalSettings, composerSendLabel,
+      activeEditorFilePath, activeEditorLineRange,
+      activeIntentCanvasCodeSelectionAnchor,
+      setActiveIntentCanvasCodeSelectionAnchor, fileReferenceMode,
+      setFileReferenceMode,
+    }),
+    panels: buildLayoutNodesPanelsOptions({
+      showComposer, activePlan, isPlanMode, openPlanPanel, closePlanPanel,
+      isPlanPanelDismissed, agentTaskScrollRequest,
+      handleSelectStatusPanelSubagent, debugEntries, debugOpen, terminalOpen,
+      terminalTabs, activeTerminalId, onSelectTerminal, onNewTerminal,
+      onCloseTerminal, terminalState, clearDebugEntries, handleCopyDebug,
+      onDebugPanelResizeStart, onTerminalPanelResizeStart, handleBackFromDiff,
+      handleGoProjects, workspaceDropTargetRef, dropOverlayActive,
+      dropOverlayText, handleWorkspaceDragOver, handleWorkspaceDragEnter,
+      handleWorkspaceDragLeave, handleWorkspaceDrop, appMode,
+      handleAppModeChange, handleOpenHomeChat, handleOpenMemory,
+      handleOpenProjectMemory, handleOpenReleaseNotes, focusedProjectMemoryId,
+      focusedProjectMemoryRequestKey, focusedWorkspaceNoteId,
+      focusedWorkspaceNoteRequestKey, handleOpenSearchPalette,
+      handleOpenQuickSwitcher, collapseSidebar, appSettings,
+      handleOpenWorkspaceHome,
+    }),
   });
 
   return {
